Task 011 - added basket domain

+++ src/main/java/com/zelinskiyrk/store/product/model/CityPriceDoc.java
@@ -11,5 +11,6 @@ import lombok.*;
 @ToString
 public class CityPriceDoc {
     private String cityId;
-    @NonNull private Integer price;
+    @NonNull
+    private Double price;
 }
